+++ data/tasks.csv
@@ -2,6 +2,6 @@ T, false, eat
 D, true, sleep, 2025-08-26
 E, false, school, 2025-09-01, 2025-09-01
 D, false, study, 2025-10-31
-T, false, shower
-D, false, shower, 2025-08-26
+D, true, shower, 2025-08-26
 E, false, shower, 2025-08-26, 2025-08-27
+D, false, wake up, 2025-08-31
