+++ src/main/java/batman/command/Command.java
@@ -1,3 +1,8 @@
+package batman.command;
+
+import batman.storage.Storage;
+import batman.task.TaskList;
+
 public abstract class Command {
     public abstract void execute(Storage storage, TaskList tasks);
 }
