+++ src/main/java/batman/command/MarkCommand.java
@@ -1,9 +1,15 @@
-public class UnmarkCommand extends Command {
+package batman.command;
+
+import batman.storage.Storage;
+import batman.task.Task;
+import batman.task.TaskList;
+
+public class MarkCommand extends Command {
     private boolean isSuccess;
     private final int index;
     private Task task;
 
-    public UnmarkCommand(int index) {
+    public MarkCommand(int index) {
         this.index = index;
         this.isSuccess = false;
     }
@@ -12,7 +18,7 @@ public class UnmarkCommand extends Command {
     public void execute(Storage storage, TaskList tasks) {
         if (this.index < tasks.getSize()) {
             this.task = tasks.getTask(index);
-            this.task.setUnmarked();
+            this.task.setMarked();
             this.isSuccess = true;
         }
     }
@@ -20,9 +26,9 @@ public class UnmarkCommand extends Command {
     @Override
     public String toString() {
         if (isSuccess) {
-            return "OK, I've marked this task as not done yet:\n" + this.task;
+            return "Nice! I've marked this task as done:\n" + this.task;
         } else {
-            return "Error: Index to be unmarked exceeds length of list";
+            return "Error: Index to be marked exceeds length of list";
         }
     }
 }
