+++ src/main/java/batman/core/Batman.class
