+++ src/main/java/batman/exception/InvalidCommandException.java
@@ -1,3 +1,5 @@
+package batman.exception;
+
 public class InvalidCommandException extends Exception {
     public InvalidCommandException() {
         super("Error: Invalid Command.");
