+++ src/main/java/batman/exception/NoDeadlineException.java
@@ -1,3 +1,5 @@
+package batman.exception;
+
 public class NoDeadlineException extends Exception {
     public NoDeadlineException() {
         super("Error: No Deadline found. Missing \"/by\" and/or deadline input in deadline task.");
