+++ src/main/java/batman/exception/NoDescriptionException.java
@@ -1,3 +1,5 @@
+package batman.exception;
+
 public class NoDescriptionException extends Exception {
     public NoDescriptionException() {
         super("Error: No Description found.");
