+++ src/main/java/batman/exception/NoFromToException.java
@@ -1,3 +1,5 @@
+package batman.exception;
+
 public class NoFromToException extends Exception {
     public NoFromToException() {
         super("Error: No event duration found. Missing \"/from\" and/or \"/to.\"");
