+++ src/main/java/batman/task/Event.java
@@ -1,3 +1,5 @@
+package batman.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
