+++ src/main/java/batman/task/TimedTask.java
@@ -1,3 +1,5 @@
+package batman.task;
+
 import java.time.format.DateTimeFormatter;
 
 public abstract class TimedTask extends Task {
